4D 1-4

+++ src/fp/vino/Vinoteca.java
@@ -32,7 +32,7 @@ public interface Vinoteca {
 	//4d
 	Map<String,List<Vino>> agruparVinosPorPais();
 	Map<String,Set<String>> agruparUvasPorPais(); 
-	Map<String,Double> calcularCalidadPrecioPorRegionMayorDe(Integer umbral);
+	Map<String,Long> calcularCalidadPrecioPorRegionMayorDe(Integer umbral);
 	Map<String,Vino> calcularVinoMasCaroPorPais();
 	SortedMap<String,List<Vino>> calcularNMejoresVinosPorPais(Integer n);
 	String calcularRegionConMejoresVinos(Integer umbral);
